# webapp/api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: VooZuu App
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/getAdminUsersList:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminUsers_ctrl
    post:
      tags: ['Admin User Module']
      description: Get Admin All Users List
      summary: Get Admin All Users List
      # used as the method name of the controller
      operationId: getAdminUsersList
      parameters:        
        - name: body
          in: body
          description: Get Admin All Users List
          schema:
           $ref: '#/definitions/getUsersList'
          required: true
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string      
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getAdminUserView/{userId}:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminUsers_ctrl
    get:
      tags: ['Admin User Module']
      description: Get User By user Id 
      summary: Get User By user Id 
      # used as the method name of the controller
      operationId: getAdminUserView
      parameters:
        - name: userId
          in: path
          description: user Id
          required: true
          type: string        
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string    
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/changeUserStatus:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminUsers_ctrl
    post:
      tags: ['Admin User Module']
      description: Update User status
      summary: Update User status
      # used as the method name of the controller
      operationId: changeUserStatus
      parameters:
        - name: body
          in: body
          description: Update User status
          schema:
           $ref: '#/definitions/changeUserStatus'
          required: true
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/bulkUpdate:
    # binds a127 app logic to a route
    x-swagger-router-controller: adminUsers_ctrl
    post:
      tags: ['Admin User Module']
      description: Update multipls Users status
      summary: Update multiple Users status
      # used as the method name of the controller
      operationId: bulkUpdate
      parameters:
        - name: body
          in: body
          description: Update multiple User status
          schema:
           $ref: '#/definitions/bulkUpdate'
          required: true
        - name: Authorization
          in: header
          description: Authorization
          required: true
          type: string        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  #Admin profile
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  changeUserStatus:   
    type: object
    properties:
      id:
        type: string
  bulkUpdate:   
    type: array
    items: 
          $ref: '#/definitions/updateMultipleUserStatus'
  updateMultipleUserStatus:
    type: object
    discriminator: type  
    properties:
      id:
       type: string
      email:
       type: string
      field:
       type: string
      status:
       type: boolean
  getUsersList:
    type: object
    properties:
      limit:
        type: number    
      page:
        type: number      
...